requirements.txt
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
et_xmlfile==2.0.0
geopandas==1.0.1
h11==0.14.0
idna==3.10
lxml==5.3.0
mysql-connector-python==9.1.0
numpy==2.2.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pycparser==2.22
pyogrio==0.10.0
pyproj==3.7.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
selenium==4.27.1
shapely==2.0.6
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
undetected-chromedriver==3.5.5
Unidecode==1.3.8
urllib3==2.3.0
websocket-client==1.8.0
websockets==14.1
wsproto==1.2.0
xlrd==2.0.1
